Refactor: move BufferSlicer into stl_util.h
